fix: Alterando DECIMAL(10,0) para BIGINT

+++ SQL/Modelo/modelo.sql
@@ -38,13 +38,13 @@ CREATE SEQUENCE public.sq_habilidade             START WITH 1 INCREMENT BY 1;
 
 
 -- Tabelas sem dependências
-CREATE TABLE public.PlanoPagamento ( nCdPlano DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_planopagamento')
+CREATE TABLE public.PlanoPagamento ( nCdPlano BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_planopagamento')
     , cNmPlano VARCHAR(50)   NOT NULL UNIQUE
     , nPreco DECIMAL(10,2)   NOT NULL
     , PRIMARY KEY(nCdPlano)
 );
 
-CREATE TABLE public.Mensagem ( nCdMensagem DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_mensagem')
+CREATE TABLE public.Mensagem ( nCdMensagem BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_mensagem')
     , cTitulo VARCHAR(50)       NOT NULL
     , cMensagem VARCHAR(255)    NOT NULL
     , cCategoria VARCHAR(70)    NOT NULL
@@ -52,21 +52,21 @@ CREATE TABLE public.Mensagem ( nCdMensagem DECIMAL(10,0) NOT NULL DEFAULT NEXTVA
 );
 
 -- Tabelas com dependência de PlanoPagamento
-CREATE TABLE public.Empresa ( nCdEmpresa        DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_empresa')
+CREATE TABLE public.Empresa ( nCdEmpresa        BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_empresa')
     , cNmEmpresa        VARCHAR(255)  NOT NULL
     , cSgEmpresa        VARCHAR(3)    NOT NULL
     , cCNPJ             VARCHAR(18)   NOT NULL UNIQUE
     , cTelefone         VARCHAR(255)  NOT NULL UNIQUE
     , cEmail            VARCHAR(255)  NOT NULL UNIQUE
     , cCEP              VARCHAR(255)  NOT NULL
-    , nCdPlanoPagamento DECIMAL(10,0) NOT NULL
+    , nCdPlanoPagamento BIGINT        NOT NULL
     , bAtivo            BOOLEAN       NOT NULL DEFAULT True
     , PRIMARY KEY (nCdEmpresa)
     , FOREIGN KEY (nCdPlanoPagamento) REFERENCES public.PlanoPagamento(nCdPlano)
 );
 
-CREATE TABLE public.PlanoVantagens ( nCdPlanoVantagem DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_planovantagens')
-    , nCdPlano         DECIMAL(10,0) NOT NULL
+CREATE TABLE public.PlanoVantagens ( nCdPlanoVantagem BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_planovantagens')
+    , nCdPlano         BIGINT        NOT NULL
     , cNmVantagem      VARCHAR(100)  NOT NULL UNIQUE
     , cDescricao       VARCHAR(255)  NOT NULL
     , PRIMARY KEY (nCdPlano, nCdPlanoVantagem)
@@ -74,16 +74,16 @@ CREATE TABLE public.PlanoVantagens ( nCdPlanoVantagem DECIMAL(10,0) NOT NULL DEF
 );
 
 -- Tabelas com dependência de Empresa
-CREATE TABLE public.Habilidade ( nCdHabilidade DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_habilidade')
-    , nCdEmpresa    DECIMAL(10,0) NOT NULL
+CREATE TABLE public.Habilidade ( nCdHabilidade BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_habilidade')
+    , nCdEmpresa    BIGINT        NOT NULL
     , cNmHabilidade VARCHAR(255)  NOT NULL
     , cDescricao    VARCHAR(255)  NOT NULL
     , PRIMARY KEY (nCdHabilidade)
     , FOREIGN KEY (nCdEmpresa) REFERENCES public.Empresa(nCdEmpresa)
 );
 
-CREATE TABLE public.Setor ( nCdSetor   DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_setor')
-    , nCdEmpresa DECIMAL(10,0) NOT NULL
+CREATE TABLE public.Setor ( nCdSetor   BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_setor')
+    , nCdEmpresa BIGINT        NOT NULL
     , cNmSetor   VARCHAR(255)  NOT NULL
     , cSgSetor   VARCHAR(3)    NOT NULL
     , PRIMARY KEY(nCdSetor)
@@ -91,12 +91,12 @@ CREATE TABLE public.Setor ( nCdSetor   DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('p
 );
 
 -- Tabela com dependência de Empresa e Setor
-CREATE TABLE public.Usuario ( nCdUsuario DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_usuario')
+CREATE TABLE public.Usuario ( nCdUsuario BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_usuario')
     , cNmUsuario  VARCHAR(100)  NOT NULL
-    , nCdGestor   DECIMAL(10,0)     NULL
+    , nCdGestor   BIGINT            NULL
     , bGestor     BOOLEAN       NOT NULL
-    , nCdEmpresa  DECIMAL(10,0) NOT NULL
-    , nCdSetor    DECIMAL(10,0) NOT NULL
+    , nCdEmpresa  BIGINT        NOT NULL
+    , nCdSetor    BIGINT        NOT NULL
     , nCPF        VARCHAR(15)   NOT NULL
     , cTelefone   VARCHAR(20)       NULL
     , cEmail      VARCHAR(255)      NULL
@@ -110,24 +110,24 @@ CREATE TABLE public.Usuario ( nCdUsuario DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL(
 );
 
 -- Tabelas com dependência de Habilidade, Usuario e Mensagem
-CREATE TABLE public.HabilidadeUsuario ( nCdHabilidade DECIMAL(10,0) NOT NULL
-    , nCdUsuario    DECIMAL(10,0) NOT NULL
+CREATE TABLE public.HabilidadeUsuario ( nCdHabilidade BIGINT        NOT NULL
+    , nCdUsuario    BIGINT        NOT NULL
     , PRIMARY KEY (nCdHabilidade, nCdUsuario)
     , FOREIGN KEY (nCdHabilidade) REFERENCES public.Habilidade (nCdHabilidade)
     , FOREIGN KEY (nCdUsuario)    REFERENCES public.Usuario (nCdUsuario)
 );
 
-CREATE TABLE public.Tarefa ( nCdTarefa         DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_tarefa')
+CREATE TABLE public.Tarefa ( nCdTarefa         BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_tarefa')
     , cNmTarefa         VARCHAR(255)  NOT NULL
-    , nCdUsuarioRelator DECIMAL(10,0) NOT NULL
-    , nCdHabilidade     DECIMAL(10,0) NOT NULL
+    , nCdUsuarioRelator BIGINT        NOT NULL
+    , nCdHabilidade     BIGINT        NOT NULL
     , iGravidade        INTEGER       NOT NULL DEFAULT 1
                                  CHECK (iGravidade >= 1 AND iGravidade <= 5)
     , iUrgencia         INTEGER       NOT NULL DEFAULT 1
                                  CHECK (iUrgencia >= 1 AND iUrgencia <= 5)
     , iTendencia        INTEGER       NOT NULL DEFAULT 1
                                  CHECK (iTendencia >= 1 AND iTendencia <= 5)
-    , nTempoEstimado    DECIMAL(10,0) NOT NULL
+    , nTempoEstimado    BIGINT        NOT NULL
     , cDescricao        VARCHAR(255)  NOT NULL
     , cStatus           VARCHAR(15)   NOT NULL DEFAULT 'Pendente'
                                  CHECK (  cStatus = 'Pendente'
@@ -141,9 +141,9 @@ CREATE TABLE public.Tarefa ( nCdTarefa         DECIMAL(10,0) NOT NULL DEFAULT NE
 );
 
 -- Tabelas com dependência de Tarefa e Usuario
-CREATE TABLE table_log.LogAtribuicaoTarefa ( nCdLogAtribuicao  DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('table_log.sq_logatribuicaotarefa')
-    , nCdTarefa         DECIMAL(10,0) NOT NULL
-    , nCdUsuarioAtuante DECIMAL(10,0) NOT NULL
+CREATE TABLE table_log.LogAtribuicaoTarefa ( nCdLogAtribuicao  BIGINT        NOT NULL DEFAULT NEXTVAL('table_log.sq_logatribuicaotarefa')
+    , nCdTarefa         BIGINT        NOT NULL
+    , nCdUsuarioAtuante BIGINT        NOT NULL
     , dRealocacao       DATE          NOT NULL DEFAULT NOW()
     , cObservacao       VARCHAR(300)  NOT NULL DEFAULT 'Nenhum comentário para a tarefa'
     , PRIMARY KEY (nCdLogAtribuicao)
@@ -151,25 +151,25 @@ CREATE TABLE table_log.LogAtribuicaoTarefa ( nCdLogAtribuicao  DECIMAL(10,0) NOT
     , FOREIGN KEY (nCdUsuarioAtuante) REFERENCES public.Usuario (nCdUsuario)
 );
 
-CREATE TABLE public.Report ( nCdReport  DECIMAL(10,0) NOT NULL DEFAULT NEXTVAL('public.sq_report')
-    , nCdTarefa  DECIMAL(10,0) NOT NULL
+CREATE TABLE public.Report ( nCdReport  BIGINT        NOT NULL DEFAULT NEXTVAL('public.sq_report')
+    , nCdTarefa  BIGINT        NOT NULL
     , cDescricao VARCHAR(255)  NOT NULL
     , cProblema  VARCHAR(255)  NOT NULL
     , PRIMARY KEY (nCdReport)
     , FOREIGN KEY (nCdTarefa) REFERENCES public.Tarefa(nCdTarefa)
 );
 
-CREATE TABLE public.TarefaHabilidade ( nCdHabilidade DECIMAL(10,0) NOT NULL
-    , nCdTarefa     DECIMAL(10,0) NOT NULL
+CREATE TABLE public.TarefaHabilidade ( nCdHabilidade BIGINT        NOT NULL
+    , nCdTarefa     BIGINT        NOT NULL
     , iPrioridade   INTEGER       NOT NULL
     , PRIMARY KEY (nCdHabilidade, nCdTarefa)
     , FOREIGN KEY (nCdHabilidade) REFERENCES public.Habilidade(nCdHabilidade)
     , FOREIGN KEY (nCdTarefa)     REFERENCES public.Tarefa (nCdTarefa)
 );
 
-CREATE TABLE public.TarefaUsuario ( nCdTarefa          DECIMAL(10,0) NOT NULL
-    , nCdUsuarioOriginal DECIMAL(10,0) NOT NULL
-    , nCdUsuarioAtuante  DECIMAL(10,0) NOT NULL
+CREATE TABLE public.TarefaUsuario ( nCdTarefa          BIGINT        NOT NULL
+    , nCdUsuarioOriginal BIGINT        NOT NULL
+    , nCdUsuarioAtuante  BIGINT        NOT NULL
     , PRIMARY KEY(nCdTarefa, nCdUsuarioOriginal, nCdUsuarioAtuante)
     , FOREIGN KEY (nCdTarefa)          REFERENCES public.Tarefa (nCdTarefa)
     , FOREIGN KEY (nCdUsuarioOriginal) REFERENCES public.Usuario(nCdUsuario)
